UAST: place method declarations together
